requirements.txt
python-socketio
flask-socketio
pybluez